+++ tests/testthat/test-afltables_player.R
@@ -11,6 +11,12 @@ test_that("get_afltables_stats works", {
   expect_error(get_afltables_stats(""))
 })
 
+test_that("get_afltables_stats reutrns the right number of rows", {
+  afltables_data_2018 <- afltables_data %>%
+    dplyr::filter(Season == 2018)
+    expect_equal(sum(afltables_data_2018$Brownlow.Votes), 1188)
+})
+
 test_that("get_afltables_stats returns correct values", {
   afltables_summary <- afltables_data %>%
     dplyr::distinct(ID, First.name, Surname) %>%