+++ NEWS.md
@@ -1,6 +1,7 @@
 # fitzRoy 0.1.12
 * Fixed an error with player ID's for 2019 season where new data was breaking an internal function ([#67](https://github.com/jimmyday12/fitzRoy/issues/67))
 * Fixed an error where Geelong v Melbourne game wasn't getting parsed properly ([#68](https://github.com/jimmyday12/fitzRoy/issues/68))
+* Fixed an error with data on github not being up to date ([#69](https://github.com/jimmyday12/fitzRoy/issues/69))
 
 # fitzRoy 0.1.11
 * Updated `get_aflw_cookie()` to align with change in site API ([#62](https://github.com/jimmyday12/fitzRoy/issues/62))